% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TbAn_Dun.R
\name{TabAnoDUN}
\alias{TabAnoDUN}
\title{Tabla ANOVA del Modelo DUNCAN.}
\usage{
TabAnoDUN(Respuesta = RESPUESTA, Tratamiento = TRATAMIENTOS, Sig = 0.05, data)
}
\arguments{
\item{Respuesta}{(vector) datos de la variable respuesta.}

\item{Tratamiento}{(vector) datos de la variable que representan el tratamientos.}

\item{Sig}{(string) variable que representa la significancia del modelo.}

\item{data}{(\code{data.frame}) Tabla de datos en formato largo con los datos de los tratamientos y de la variable respuesta.}
}
\value{
Devuelve una tabla en formato \code{data.frame} con los cálculos correspondientes al análisis de varianza que compete a la Prueba de DUNCAN.
}
\description{
Obtiene la Tabla de Análisis de Varianza (ANOVA) para la Prueba de Rangos Multiples DUNCAN (PRMD).
}
