% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Med_Dun.R
\name{TabDUN}
\alias{TabDUN}
\title{Tabla de Separación de Medias del Modelo DUNCAN.}
\usage{
TabDUN(Respuesta, Tratamiento, s, rp05, rp01, data)
}
\arguments{
\item{Respuesta}{(vector) datos de la variable respuesta.}

\item{Tratamiento}{(vector) datos de la variable que representan el tratamientos.}

\item{s}{(string) variable que representa la significancia del modelo.}

\item{rp05}{(string) variable que respresenta la razon que deriva de una posibilidad al 5% con ciertos grados de libertad.}

\item{rp01}{(string) variable que respresenta la razon que deriva de una posibilidad al 1% con ciertos grados de libertad.}

\item{data}{(\code{data.frame}) Tabla de datos en formato largo con los datos de los tratamientos y de la variable respuesta.}
}
\value{
Devuelve una tabla en formato \code{data.frame} con los cálculos correspondientes a la diferencia entre medias de las variables respuesta por tratamiento que compete a la Prueba de DUNCAN.
}
\description{
Obtiene la Tabla de Comparacion de Medias (Diferencia de medias) para la Prueba de Rangos Multiples DUNCAN (PRMD).
}
